+++ cmd/ttn-lw-cli/commands/gateways.go
@@ -16,8 +16,11 @@ package commands
 
 import (
 	"os"
+	"strings"
 
+	"github.com/TheThingsIndustries/protoc-gen-go-flags/flagsplugin"
 	"github.com/gogo/protobuf/proto"
+	"github.com/gogo/protobuf/types"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"go.thethings.network/lorawan-stack/v3/cmd/internal/io"
@@ -432,13 +435,28 @@ var (
 		},
 	}
 	gatewaysConnectionStats = &cobra.Command{
-		Use:     "get-connection-stats [gateway-id]",
+		Use:     "get-connection-stats [gateway-ids]",
 		Aliases: []string{"connection-stats", "cnx-stats", "stats"},
-		Short:   "Get connection stats for a gateway",
+		Short:   "Get connection stats for a (group of) gateway(s).",
+		Long: `Get connection stats for a (group of) gateway(s).
+The command arguments support multiple gateway IDs.
+Alternatively use the --gateway-ids flag.
+If both the parameter and the flag are provided, the flag is ignored.`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			gtwID, err := getGatewayID(cmd.Flags(), args, true)
-			if err != nil {
-				return err
+			var (
+				ids = args
+				err error
+			)
+			if len(ids) == 0 {
+				raw, err := cmd.Flags().GetString("gateway-ids")
+				if err != nil {
+					return err
+				}
+				s := strings.Split(raw, ",")
+				if len(s) < 2 {
+					return errNoGatewayID.New()
+				}
+				ids = append(ids, s...)
 			}
 
 			is, err := api.Dial(ctx, config.IdentityServerGRPCAddress)
@@ -446,28 +464,52 @@ var (
 				return err
 			}
 
-			gateway, err := ttnpb.NewGatewayRegistryClient(is).Get(ctx, &ttnpb.GetGatewayRequest{
-				GatewayIds: gtwID,
-				FieldMask:  ttnpb.FieldMask("gateway_server_address"),
-			})
-			if err != nil {
-				return err
-			}
-
-			if gsMismatch := compareServerAddressGateway(gateway, config); gsMismatch {
-				return errAddressMismatchGateway.New()
-			}
-
 			gs, err := api.Dial(ctx, config.GatewayServerGRPCAddress)
 			if err != nil {
 				return err
 			}
 
-			res, err := ttnpb.NewGsClient(gs).GetGatewayConnectionStats(ctx, gtwID)
+			if len(ids) > 1 {
+				paths, err := cmd.Flags().GetStringSlice("paths")
+				if err != nil {
+					return err
+				}
+
+				var gtwIDs []*ttnpb.GatewayIdentifiers
+				for _, id := range ids {
+					gtwIDs = append(gtwIDs, &ttnpb.GatewayIdentifiers{
+						GatewayId: id,
+					})
+				}
+				res, err := ttnpb.NewGsClient(gs).BatchGetGatewayConnectionStats(ctx,
+					&ttnpb.BatchGetGatewayConnectionStatsRequest{
+						GatewayIds: gtwIDs,
+						FieldMask: &types.FieldMask{
+							Paths: paths,
+						},
+					})
+				if err != nil {
+					return err
+				}
+				return io.Write(os.Stdout, config.OutputFormat, res)
+			}
+			gtwIDs := &ttnpb.GatewayIdentifiers{
+				GatewayId: ids[0],
+			}
+			gateway, err := ttnpb.NewGatewayRegistryClient(is).Get(ctx, &ttnpb.GetGatewayRequest{
+				GatewayIds: gtwIDs,
+				FieldMask:  ttnpb.FieldMask("gateway_server_address"),
+			})
 			if err != nil {
 				return err
 			}
-
+			if gsMismatch := compareServerAddressGateway(gateway, config); gsMismatch {
+				return errAddressMismatchGateway.New()
+			}
+			res, err := ttnpb.NewGsClient(gs).GetGatewayConnectionStats(ctx, gtwIDs)
+			if err != nil {
+				return err
+			}
 			return io.Write(os.Stdout, config.OutputFormat, res)
 		},
 	}
@@ -547,6 +589,15 @@ func init() {
 	gatewaysRestoreCommand.Flags().AddFlagSet(gatewayIDFlags())
 	gatewaysCommand.AddCommand(gatewaysRestoreCommand)
 	gatewaysConnectionStats.Flags().AddFlagSet(gatewayIDFlags())
+	gatewaysConnectionStats.Flags().StringSlice(
+		"paths",
+		[]string{},
+		"comma separated list of paths to filter on a batch of stats")
+	flagsplugin.AddAlias(
+		gatewaysConnectionStats.Flags(),
+		"gateway-id",
+		"gateway-ids",
+	)
 	gatewaysCommand.AddCommand(gatewaysConnectionStats)
 	gatewaysContactInfoCommand.PersistentFlags().AddFlagSet(gatewayIDFlags())
 	gatewaysCommand.AddCommand(gatewaysContactInfoCommand)