+++ pkg/gatewayserver/redis/registry.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/go-redis/redis/v8"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
+	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	ttnredis "go.thethings.network/lorawan-stack/v3/pkg/redis"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/unique"
@@ -85,4 +86,63 @@ func (r *GatewayConnectionStatsRegistry) Get(ctx context.Context, ids *ttnpb.Gat
 		return nil, ttnredis.ConvertError(err)
 	}
 	return result, nil
+}
+
+func applyGatewayConnectionStatsFieldMask(
+	dst, src *ttnpb.GatewayConnectionStats,
+	paths ...string,
+) (*ttnpb.GatewayConnectionStats, error) {
+	if dst == nil {
+		dst = &ttnpb.GatewayConnectionStats{}
+	}
+	return dst, dst.SetFields(src, paths...)
+}
+
+// BatchGet returns the connection stats for a batch of gateways.
+// NotFound errors indicating that the gateway is either not connected
+// or is connected to a different cluster, are ignored.
+func (r *GatewayConnectionStatsRegistry) BatchGet(
+	ctx context.Context,
+	ids []*ttnpb.GatewayIdentifiers,
+	paths []string,
+) (map[string]*ttnpb.GatewayConnectionStats, error) {
+	ret := make(map[string]*ttnpb.GatewayConnectionStats, len(ids))
+	keys := make([]string, 0, len(ids))
+	for _, gtwIDs := range ids {
+		uid := unique.ID(ctx, gtwIDs)
+		keys = append(keys, r.key(uid))
+	}
+	rawValues, err := r.Redis.MGet(ctx, keys...).Result()
+	if err != nil {
+		return nil, ttnredis.ConvertError(err)
+	}
+
+	for i, val := range rawValues {
+		switch val := val.(type) {
+		case nil:
+			continue
+		case string:
+			stats := &ttnpb.GatewayConnectionStats{}
+			if err := ttnredis.UnmarshalProto(val, stats); err != nil {
+				log.FromContext(ctx).WithError(err).Warnf("Failed to decode stats payload")
+				continue
+			}
+			// Copy only the requested paths.
+			if len(paths) > 0 {
+				stats, err = applyGatewayConnectionStatsFieldMask(nil, stats, paths...)
+				if err != nil {
+					return nil, err
+				}
+			}
+
+			// The result of MGet is in the same order as the input keys passed to it.
+			// MGet inserts "nil" values for keys that don't have values, thereby maintaining the order.
+			// So we can use the index of the result to correlate the gateway IDs.
+			ret[ids[i].GatewayId] = stats
+		default:
+			log.FromContext(ctx).WithField("element", val).Warn("Invalid element in stats payloads")
+			continue
+		}
+	}
+	return ret, nil
 }