+++ pkg/ttnpb/gatewayserver.pb.paths.fm.go
@@ -277,3 +277,19 @@ var ScheduleDownlinkErrorDetailsFieldPathsNested = []string{
 var ScheduleDownlinkErrorDetailsFieldPathsTopLevel = []string{
 	"path_errors",
 }
+var BatchGetGatewayConnectionStatsRequestFieldPathsNested = []string{
+	"field_mask",
+	"gateway_ids",
+}
+
+var BatchGetGatewayConnectionStatsRequestFieldPathsTopLevel = []string{
+	"field_mask",
+	"gateway_ids",
+}
+var BatchGetGatewayConnectionStatsResponseFieldPathsNested = []string{
+	"entries",
+}
+
+var BatchGetGatewayConnectionStatsResponseFieldPathsTopLevel = []string{
+	"entries",
+}